+++ changes/issue-7080-new-query-sidebar-docs
@@ -1,0 +1,1 @@
+- add new query sidebar with updated and improved docs
