+++ frontend/components/FleetMarkdown/_styles.scss
@@ -6,4 +6,16 @@
     // the container.
     padding-left: 20px;
   }
+
+  pre {
+    padding: 12px;
+    border: 1px solid $ui-blue-gray;
+    background-color: $ui-light-grey; // copy fleet ace background color
+
+    .ace_cursor {
+        // We have the !important here as there doesnt seen a way to programatically
+        // hide only the cursor in the editor.
+        display: none !important
+      }
+  }
 }
